requirements-dev.txt
psycopg2-binary>=2.8
djangorestframework>=3.13
django-environ>=0.8
django-filter>=21.1
django-guardian>=2.4.0
djangorestframework-gis>=0.18
redis>=4.0
hiredis>=2.0
requests>=2.27
django-storages[google]>=1.12
pillow>=9.0
channels>=3.0
channels_redis>=3.4
selenium>=4.1.3
celery[redis]>=5.2.3
pyparsing>=3.0.8
Werkzeug>=2.1.1
pygments>=2.12
pystac>=1.6.1
environ>=1.0
daphne>=3.0.2
django-cors-headers>=3.8.0
django-debug-toolbar==3.7.0
django-extensions==3.2
actinia-python-client @ https://github.com/actinia-org/actinia-python-client/releases/download/0.3.0/actinia_python_client-0.3.0-py3-none-any.whl
requests-mock==1.11
